chore: refactoring carbon export code
